chore: refactor out VecTextBuffer tests to a common test set
