+++ crates/knap_ui/src/text_buffer/vec_text_buffer.rs
@@ -235,343 +235,9 @@ mod tests {
     }
 
     #[test]
-    fn test_line() {
-        // normal
-        {
-            let buffer = VecTextBuffer {
-                text: vec![
-                    "Hello".to_string(),
-                    "World!".to_string(),
-                    "".to_string(),
-                    "The End".to_string(),
-                ],
-            };
+    fn test_standard_text_buffer_tests() {
+        use crate::text_buffer::buffer_tests::do_standard_text_buffer_tests;
 
-            assert_eq!(buffer.line(0), Some("Hello".to_string()));
-            assert_eq!(buffer.line(1), Some("World!".to_string()));
-            assert_eq!(buffer.line(2), Some("".to_string()));
-            assert_eq!(buffer.line(3), Some("The End".to_string()));
-            assert_eq!(buffer.line(4), None);
-        }
-
-        // empty
-        {
-            let buffer = VecTextBuffer::new();
-            assert_eq!(buffer.line(0), None);
-        }
-    }
-
-    #[test]
-    fn test_line_len() {
-        // normal
-        {
-            let buffer = VecTextBuffer {
-                text: vec![
-                    "Hello".to_string(),
-                    "World!".to_string(),
-                    "".to_string(),
-                    "The End".to_string(),
-                ],
-            };
-
-            assert_eq!(buffer.line_len(0), Some(5));
-            assert_eq!(buffer.line_len(1), Some(6));
-            assert_eq!(buffer.line_len(2), Some(0));
-            assert_eq!(buffer.line_len(3), Some(7));
-            assert_eq!(buffer.line_len(4), None);
-        }
-
-        // empty
-        {
-            let buffer = VecTextBuffer::new();
-            assert_eq!(buffer.line_len(0), None);
-        }
-    }
-
-    #[test]
-    fn test_total_lines() {
-        // normal
-        {
-            let buffer = VecTextBuffer {
-                text: vec![
-                    "Hello".to_string(),
-                    "World!".to_string(),
-                    "".to_string(),
-                    "The End".to_string(),
-                ],
-            };
-
-            assert_eq!(buffer.total_lines(), 4);
-        }
-
-        // empty
-        {
-            let buffer = VecTextBuffer::new();
-            assert_eq!(buffer.total_lines(), 0);
-        }
-    }
-
-    #[test]
-    fn test_insert_character_at_pos() {
-        let mut buffer = VecTextBuffer::new();
-
-        // insert in empty buffer
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 0, byte: 0 }, 'a');
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "a");
-
-        // insert in non-empty buffer
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 0, byte: 1 }, 'b');
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "ab");
-
-        // insert at the end of the line
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 0, byte: 2 }, 'c');
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "abc");
-
-        // insert at the front of the line
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 0, byte: 0 }, 'd');
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "dabc");
-
-        // insert in the middle of the line
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 0, byte: 2 }, 'e');
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "daebc");
-        assert_eq!(buffer.total_lines(), 1);
-
-        // insert newline character, causes a new line to be created
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 0, byte: 5 }, '\n');
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "daebc\n");
-        assert_eq!(buffer.total_lines(), 2);
-
-        // can insert character in the new line
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 1, byte: 0 }, 'f');
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "daebc\nf");
-
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 1, byte: 1 }, 'g');
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "daebc\nfg");
-
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 1, byte: 1 }, 'h');
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "daebc\nfhg");
-
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 1, byte: 0 }, 'i');
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "daebc\nifhg");
-
-        // can still insert at old line
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 0, byte: 5 }, 'j');
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "daebcj\nifhg");
-
-        // fail to insert beyond the end of the line
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 0, byte: 7 }, 'z');
-        assert_eq!(result, Err(InsertCharError::InvalidBytePosition));
-        assert_eq!(buffer.contents(), "daebcj\nifhg");
-
-        // fail to insert on a non-existent line that is not directly after the last line
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 3, byte: 0 }, 'z');
-        assert_eq!(result, Err(InsertCharError::InvalidLinePosition));
-        assert_eq!(buffer.contents(), "daebcj\nifhg");
-
-        // ok to insert at byte 0 after the last line. a newline should also be automatically inserted
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 2, byte: 0 }, 'k');
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "daebcj\nifhg\nk");
-        assert_eq!(buffer.total_lines(), 3);
-
-        // but if try to insert at byte non-0 after the last line, it will fail
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 3, byte: 1 }, 'z');
-        assert_eq!(result, Err(InsertCharError::InvalidBytePosition));
-        assert_eq!(buffer.contents(), "daebcj\nifhg\nk");
-        assert_eq!(buffer.total_lines(), 3);
-
-        // insert newline in the middle of a line causes the line to be split
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 0, byte: 2 }, '\n');
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "da\nebcj\nifhg\nk");
-        assert_eq!(buffer.total_lines(), 4);
-
-        // insert newline at the start of the line spawns a line before
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 2, byte: 0 }, '\n');
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "da\nebcj\n\nifhg\nk");
-        assert_eq!(buffer.total_lines(), 5);
-
-        // insert newline at the end of the line spawns a line below
-        let result = buffer.insert_character_at_pos(TextBufferPos { line: 2, byte: 0 }, '\n');
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "da\nebcj\n\n\nifhg\nk");
-        assert_eq!(buffer.total_lines(), 6);
-    }
-
-    #[test]
-    fn test_remove_character_at_pos() {
-        let mut buffer = VecTextBuffer::new();
-        buffer.set_contents("Hello\nWorld!\nAnother\nLine");
-
-        // delete first character of line
-        let result = buffer.remove_character_at_pos(TextBufferPos { line: 2, byte: 0 });
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "Hello\nWorld!\nnother\nLine");
-
-        // delete last character of line
-        let result = buffer.remove_character_at_pos(TextBufferPos { line: 2, byte: 5 });
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "Hello\nWorld!\nnothe\nLine");
-
-        // delete character in the middle of the line
-        let result = buffer.remove_character_at_pos(TextBufferPos { line: 2, byte: 2 });
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "Hello\nWorld!\nnohe\nLine");
-
-        // delete newline character, resulting in a join
-        let result = buffer.remove_character_at_pos(TextBufferPos { line: 0, byte: 5 });
-        assert_eq!(result, Ok(()));
-        assert_eq!(buffer.contents(), "HelloWorld!\nnohe\nLine");
-
-        // fail to delete beyond last line of character (including newline)
-        let result = buffer.remove_character_at_pos(TextBufferPos { line: 0, byte: 12 });
-        assert_eq!(result, Err(RemoveCharError::InvalidBytePosition));
-        assert_eq!(buffer.contents(), "HelloWorld!\nnohe\nLine");
-
-        // fail to delete non-existent line
-        let result = buffer.remove_character_at_pos(TextBufferPos { line: 3, byte: 0 });
-        assert_eq!(result, Err(RemoveCharError::InvalidLinePosition));
-        assert_eq!(buffer.contents(), "HelloWorld!\nnohe\nLine");
-
-        // fail to delete in empty buffer
-        buffer.set_contents("");
-        let result = buffer.remove_character_at_pos(TextBufferPos { line: 0, byte: 0 });
-        assert_eq!(result, Err(RemoveCharError::InvalidLinePosition));
-        assert_eq!(buffer.contents(), "");
-    }
-
-    #[test]
-    fn test_find() {
-        let mut buffer = VecTextBuffer::new();
-        buffer.set_contents("this is a text and this is his cat.\nthe next line contains the history.\nand this is the last line.");
-
-        // normal search on the same line
-        let result = buffer.find(
-            "this",
-            TextBufferPos { line: 0, byte: 0 },
-            SearchDirection::Forward,
-        );
-        assert_eq!(result, Some(TextBufferPos { line: 0, byte: 0 }));
-
-        let result = buffer.find(
-            "this",
-            TextBufferPos { line: 0, byte: 1 },
-            SearchDirection::Forward,
-        );
-        assert_eq!(result, Some(TextBufferPos { line: 0, byte: 19 }));
-
-        let result = buffer.find(
-            "this",
-            TextBufferPos { line: 0, byte: 18 },
-            SearchDirection::Forward,
-        );
-        assert_eq!(result, Some(TextBufferPos { line: 0, byte: 19 }));
-
-        let result = buffer.find(
-            "this",
-            TextBufferPos { line: 0, byte: 19 },
-            SearchDirection::Forward,
-        );
-        assert_eq!(result, Some(TextBufferPos { line: 0, byte: 19 }));
-
-        let result = buffer.find(
-            "this",
-            TextBufferPos { line: 0, byte: 21 },
-            SearchDirection::Backward,
-        );
-        assert_eq!(result, Some(TextBufferPos { line: 0, byte: 19 }));
-
-        let result = buffer.find(
-            "this",
-            TextBufferPos { line: 0, byte: 20 },
-            SearchDirection::Backward,
-        );
-        assert_eq!(result, Some(TextBufferPos { line: 0, byte: 19 }));
-
-        let result = buffer.find(
-            "this",
-            TextBufferPos { line: 0, byte: 19 },
-            SearchDirection::Backward,
-        );
-        assert_eq!(result, Some(TextBufferPos { line: 0, byte: 0 }));
-
-        let result = buffer.find(
-            "this",
-            TextBufferPos { line: 0, byte: 18 },
-            SearchDirection::Backward,
-        );
-        assert_eq!(result, Some(TextBufferPos { line: 0, byte: 0 }));
-
-        // jump to a different line
-        let result = buffer.find(
-            "this",
-            TextBufferPos { line: 0, byte: 20 },
-            SearchDirection::Forward,
-        );
-        assert_eq!(result, Some(TextBufferPos { line: 2, byte: 4 }));
-
-        let result = buffer.find(
-            "this",
-            TextBufferPos { line: 2, byte: 4 },
-            SearchDirection::Backward,
-        );
-        assert_eq!(result, Some(TextBufferPos { line: 0, byte: 19 }));
-
-        // wrap around
-        let result = buffer.find(
-            "is",
-            TextBufferPos { line: 2, byte: 10 },
-            SearchDirection::Forward,
-        );
-        assert_eq!(result, Some(TextBufferPos { line: 0, byte: 2 }));
-
-        let result = buffer.find(
-            "is",
-            TextBufferPos { line: 0, byte: 2 },
-            SearchDirection::Backward,
-        );
-        assert_eq!(result, Some(TextBufferPos { line: 2, byte: 9 }));
-
-        // non-existent search
-        let result = buffer.find(
-            "cannot be found",
-            TextBufferPos { line: 0, byte: 0 },
-            SearchDirection::Forward,
-        );
-        assert_eq!(result, None);
-
-        let result = buffer.find(
-            "cannot be found",
-            TextBufferPos { line: 2, byte: 5 },
-            SearchDirection::Backward,
-        );
-        assert_eq!(result, None);
-
-        // search in empty buffer
-        let empty_buffer = VecTextBuffer::new();
-        let result = empty_buffer.find(
-            "nothing to search",
-            TextBufferPos { line: 0, byte: 0 },
-            SearchDirection::Forward,
-        );
-        assert_eq!(result, None);
-        let result = empty_buffer.find(
-            "nothing to search",
-            TextBufferPos { line: 0, byte: 0 },
-            SearchDirection::Backward,
-        );
-        assert_eq!(result, None);
+        do_standard_text_buffer_tests(&|| VecTextBuffer::new());
     }
 }
